move read_pdf to pdf_parser.py
